Fix deleting an uploaded file
If we delete an uploaded file, the entity is retrieved from the database and its file field is therefore not an instance of UploadedFile, but a string. With this change, we check whether the file field is an instance of UploadedFile or a string, and ensure that the file name is correctly set.